+++ cameo-api-cpp/include/Services.h
@@ -51,7 +51,8 @@ public:
 	void retrieveServerVersion();
 	void initStatus();
 	std::unique_ptr<EventStreamSocket> openEventStream();
-	std::unique_ptr<OutputStreamSocket> createOutputStreamSocket(int port);
+	int getStreamPort(const std::string& name);
+	std::unique_ptr<OutputStreamSocket> createOutputStreamSocket(const std::string& name);
 	std::unique_ptr<RequestSocketImpl> createRequestSocket(const std::string& endpoint);
 	std::unique_ptr<RequestSocketImpl> createRequestSocket(const std::string& endpoint, int timeout);